+++ src/main/java/com/bookbla/americano/scheduler/ScheduleWorker.java
@@ -9,6 +9,7 @@ import com.bookbla.americano.domain.member.repository.MemberRepository;
 import com.bookbla.americano.domain.member.repository.entity.MemberBookmark;
 import com.bookbla.americano.domain.notification.service.MailService;
 import com.bookbla.americano.domain.payment.infrastructure.google.GoogleCertificationProvider;
+import com.bookbla.americano.domain.payment.service.GooglePaymentService;
 import com.bookbla.americano.domain.postcard.enums.PostcardStatus;
 import com.bookbla.americano.domain.postcard.repository.PostcardRepository;
 import com.bookbla.americano.domain.postcard.repository.entity.Postcard;
@@ -40,7 +41,7 @@ class ScheduleWorker {
     private final BookblaLogDiscord bookblaLogDiscord;
     private final MemberBookmarkRepository memberBookmarkRepository;
     private final PostcardRepository postcardRepository;
-    private final GoogleCertificationProvider googleCertificationProvider;
+    private final GooglePaymentService googlePaymentService;
 
     @Scheduled(cron = EVERY_4_AM, zone = "Asia/Seoul")
     public void deleteMemberEmailSchedule() {
@@ -142,7 +143,7 @@ class ScheduleWorker {
     @Scheduled(cron = EVERY_0_AM, zone = "Asia/Seoul")
     public void refundVoidedPurchaseSchedule() {
         try {
-            googleCertificationProvider.refundVoidedPurchase(0, 50);
+            googlePaymentService.refundVoidedPurchase(0, 50);
         } catch (Exception e) {
             String txName = ScheduleWorker.class.getName() + "(refundVoidedPurchaseSchedule)";
             String message = "구글 인앱 결제 환불 처리 작업이 실패하였습니다. 확인 부탁드립니다." + CRLF
